% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcgam.R
\name{as.gam}
\alias{as.gam}
\title{Convert an rcgam to a gam
Useful for applying gam-specific methods that may not work with e.g. predict.rcgam}
\usage{
as.gam(object)
}
\description{
Convert an rcgam to a gam
Useful for applying gam-specific methods that may not work with e.g. predict.rcgam
}

