% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeModelData.R
\name{calcLoad}
\alias{calcLoad}
\title{Calculate loads from concentration and flow}
\usage{
calcLoad(flow, conc, flow.units = "CFS", conc.units = "mg/l",
  load.units = "kg/day")
}
\description{
Performs appropriate unit conversion to go from flow and concentration to load
}

