% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condlSample.R
\name{condlSample.rcgam}
\alias{condlSample.rcgam}
\alias{condlSample.rclm}
\title{Predict method for rclm/ rcgam fits}
\usage{
\method{condlSample}{rcgam}(object, newdata, flowcol = "flow",
  flow.units = "CFS", quantile, ...)

\method{condlSample}{rclm}(object, newdata, flowcol = "flow",
  flow.units = "CFS", quantile, ...)
}
\arguments{
\item{object}{An rclm or rcgam object to use for predicting}

\item{newdata}{a data.frame containing precictor variables to use for prediction}

\item{...}{Arguments passed to `predict.lm` or predict.gam function call}

\item{retransform}{Should the predictions be returned as concentrations? (defaults to TRUE)}

\item{smear}{Use Smearing estimator to correct transformation bias?}

\item{object}{An rclm or rcgam object to use for predicting}

\item{newdata}{a data.frame containing precictor variables to use for prediction}

\item{retransform}{Should the predictions be returned as concentrations? (defaults to TRUE)}

\item{...}{Arguments passed to `predict.lm` or predict.gam function call}

\item{smear}{Use Smearing estimator to correct transformation bias?}
}
\description{
Predict values using an rclm or rcgam model object

Predict values using an rclm or rcgam model object
}
\section{Methods (by class)}{
\itemize{
\item \code{rcgam}: Method for rcgam objects

\item \code{rclm}: Method for rclm objects
}}

