% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{convertUnits}
\alias{convertUnits}
\title{Wrapper for ud.convert that accepts vector inputs and multiple conversion candidates.}
\usage{
convertUnits(x, from, to)
}
\arguments{
\item{x}{Numeric vector of values}

\item{from}{Units of x values}

\item{to}{Units to convert to.}
}
\value{
A data.frame with columns x and units
}
\description{
Wrapper for ud.convert that accepts vector inputs and multiple conversion candidates.
}
\details{
If length(to) > 0, units the first is tried first, followed by the second, etc.
 If units in from are already contained in to, these units are not converted.
}

