% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{discount}
\alias{discount}
\title{discount factor described in Wang et al., 2011}
\usage{
discount(x, d)
}
\arguments{
\item{x}{timeseries to exponentially smooth}

\item{d}{parameter of smooth. Must be strictly between 0 and 1. See Wang et al., 2011}
}
\description{
Applies an exponential smooth function with parameter p
}

