% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModelData.R
\name{getData.rclm}
\alias{getData.rclm}
\title{Get data from rcgam or rclm objects}
\usage{
\method{getData}{rclm}(object, type = c("raw", "rcData"))
}
\arguments{
\item{object}{an object of class `rclm`}

\item{type}{What kind of data to return--raw or transformed (rcData object)}
}
\description{
simple extraction of data, returning useful errors if impossible.
}

