% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharm.R, R/utils.R
\name{jsec}
\alias{jsec}
\title{obtain "julian seconds" (number of seconds since start of the year)}
\usage{
jsec(t)

jsec(t)
}
\description{
obtain "julian seconds" (number of seconds since start of the year)

obtain "julian seconds" (number of seconds since start of the year)
}

