% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadest.R
\name{loadest_cal}
\alias{loadest_cal}
\title{Calibrate a model or series of models using LOADEST built-in forms}
\usage{
loadest_cal(data, modno = 0, fun.select = AIC, which.select = min)
}
\arguments{
\item{data}{An rcData object}

\item{modno}{The LOADEST model number, defaults to 0 (automatic selection)}

\item{fun.select}{function to use to select between models}

\item{which.select}{Either min or max, tells which value of select funciton is best}
}
\description{
Returns list with elements corresponding to typical LOADEST output files
}

