% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadest.R
\name{loadest_pred}
\alias{loadest_pred}
\title{Make predictions a la LOADEST}
\usage{
loadest_pred(object, preddata, flow = NULL, what = c("load", "conc"), ...)
}
\arguments{
\item{object}{a rcgam object}

\item{preddata}{either a rcData object or a data.frame that can be converted to one via makePredData}

\item{what}{either "load" for load (kg/day) or "conc" for concentration (mg/L)}
}
\description{
Make predictions a la LOADEST
}

