% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLoad.R
\name{loadTS}
\alias{loadTS}
\title{Calculate loads from concentration and flow}
\usage{
loadTS(flow, conc, datetime, flow.units = "ft3/s", conc.units = "mg/l",
  load.units = "kg/day")
}
\arguments{
\item{flow}{stream discharge (volume per timie)}

\item{conc}{constituent concentration (mass per volume)}

\item{datetime}{Time of observations, as a Date or POSIXct object.}

\item{flow.units}{options are "ft3/s" (default), "m3/s", and "l/s"}

\item{conc.units}{options are "mg/l" (default), "ug/l", and "ng/l"}

\item{load.units}{options are "kg/day"}
}
\value{
data.frame with columns "datetime" and "load", also attribute "load.units"
}
\description{
Performs appropriate unit conversion to go from flow and concentration to load
}

