% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeModelData.R
\name{makePredData}
\alias{makePredData}
\title{Make prediction data from "raw" data}
\usage{
makePredData(rawData, object)
}
\arguments{
\item{rawData}{The data to convert. Currently methods only exist for
data.frame, but rcData objects may be added in the future.}

\item{model}{A model from which to extract transformation
information.}
}
\description{
Similar to makeModelData, but requires a model to be supplied and does
not expect concentration data to be present.
}

