% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeModelData.R
\name{makeRawData}
\alias{makeRawData}
\title{Make raw data frame from a rcData object}
\usage{
makeRawData(data, ...)
}
\arguments{
\item{data}{transformed data.}

\item{tfm}{a list containing transformation functions, not necessary if data is an rcData object.}
}
\description{
Generic function for getting untransformed data from transformed data and
some specification of transformation.
}

