% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.R
\name{makeSyntheticCohn}
\alias{makeSyntheticCohn}
\title{Make a synthetic linear dataset
Based on description of Monte Carlo experiments in Cohn, 2005.}
\usage{
makeSyntheticCohn(nrows, sigsq, log_detlim = 0)
}
\arguments{
\item{nrows}{The number of sumulations}

\item{sigsq}{the sigma^2 parameter. Cohn (2005) used 0.04, 0.25, 0.64, and 1.44.}

\item{log_detlim}{the imposed detection limit for censoring the observations.
Cohn (2005) used -Inf, -2*sigma, -sigma, 0, sigma}
}
\value{
A data.frame with columns "Date" (arbitrarily simulated),
"conc", "flow", "conc.units", "flow.units", and "is.bdl".
}
\description{
Make a synthetic linear dataset
Based on description of Monte Carlo experiments in Cohn, 2005.
}

