% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{NSE}
\alias{NSE}
\title{Nash-sutcliffe efficiency for concentration and load}
\usage{
NSE(object, what = c("concentration", "load"))
}
\arguments{
\item{object}{a rcgam or rclm object}

\item{what}{Calculate NSE on concentration or load?}
}
\description{
Nash-sutcliffe efficiency for concentration and load
}

