% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcgam.R
\name{predict.rcgam}
\alias{predict.rcgam}
\title{Predict method for rcgam model objects}
\usage{
\method{predict}{rcgam}(object, smear = TRUE, retrans = TRUE, ...)
}
\value{
a list with element "fit" and potentially others,
depending on optional arguments passed. See help("predict.gam", package = "mgcv") for more info
The difference is that here the returned value will always be a list, possibly of length 1.
}
\description{
Predict method for rcgam model objects
}

