% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rclm.R
\name{predict.rclm}
\alias{predict.rclm}
\title{Predict method for rclm fits}
\usage{
\method{predict}{rclm}(object, newdata, ..., smear = TRUE,
  flowcol = attr(object, "flowcol"), ccol = attr(object, "ccol"),
  datecol = attr(object, "datecol"), logflowvar = attr(object,
  "logflowvar"), logcvar = attr(object, "logcvar"), doyvar = attr(object,
  "doyvar"), numtimevar = attr(object, "numtimevar"))
}
\arguments{
\item{object}{An rclm object to use for predicting}

\item{newdata}{a data.frame containing precictor variables to use for prediction}

\item{...}{Arguments passed to `predict.lm` function call}

\item{smear}{Use Smearing estimator to correct transformation bias?}
}
\description{
Predict values using an rclm model object
}

