% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcgam.R
\name{rcgam}
\alias{rcgam}
\title{Rating-curve generalized additive model}
\usage{
rcgam(formula, data, timeout = 1, ...)
}
\arguments{
\item{formula}{a GAM formula. See `help("gam", package = "mgcv")` for details.}

\item{data}{an object of class rcData, generated using `makeModelData`, or a data frame that makeModelData can use.}

\item{timeout}{the amount of time in seconds to try mgcv::gam before stopping with a timeout error}

\item{...}{further arguments passed to `mgcv::gam`}
}
\description{
Fits generalized additive models to log-log relationships using `mgcv::gam`,
but incorporates aspects
specific to hydrologic rating curves. Specifically, tracks the geometric mean of
flow and concentration, adjusts for transformation bias, and calculates relevant
fit statistics such as Nash-Sutcliffe efficiency (NSE)
}
\details{
Dots get passed to mgcv::gam() call.
All logarithms are base e (natural log)
Arguments ending in "var" are names of objects that will be assigned in the function.
logflowvar and logcvar will be shifted to have mean zero. POSSIBLY CHANGE THIS?
}

