% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclm.R
\name{rclm}
\alias{rclm}
\title{Linear Rating-curve model}
\usage{
rclm(formula, data, ...)
}
\arguments{
\item{formula}{a lm formula. See `help("lm")` for details.}

\item{data}{an object of class rcData, generated using `makeModelData`, or a data frame that makeModelData can use.}

\item{...}{further arguments passed to `mgcv::gam`}

\item{timeout}{the amount of time in seconds to try mgcv::gam before stopping with a timeout error}
}
\description{
Fits linear models to log-log relationships using `mgcv::lm`,
but incorporates aspects
specific to hydrologic rating curves. Specifically, tracks the geometric mean of
flow and concentration, adjusts for transformation bias, and calculates relevant
fit statistics such as Nash-Sutcliffe efficiency (NSE)
}
\details{
Dots get passed to lm() call.
All logarithms are base e (natural log)
Arguments ending in "var" are names of objects that will be assigned in the function.
logflowvar and logcvar will be shifted to have mean zero. POSSIBLY CHANGE THIS?
}

