% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sharm.R
\name{sharm}
\alias{sharm}
\title{Compute seasonal harmonic functions}
\usage{
sharm(t, degree = 1)
}
\arguments{
\item{t}{a Date or POSIXt vector at which to evaluate the harmonic series.}

\item{degree}{the degree of the highest-degree harmonic function}
}
\value{
An object of class "sharm", which is a matrix with a `degree` attribute
corresponding to the maximum degree of the harmonic series.
}
\description{
Returns orthogonal harmonic functions from degree 1 to `degree` over a
specified set of dates, such that all have a common period of 1 year.
Based on the `stats::poly` function.
}
\details{
returns a matrix with (2 * degree) columns, representing the powers-of-2 multiples of
sin and cosine functions with 1-year period.

Useful for adding seasonal harmonic terms in regression functions.
Uses 1 January as origin. Incorporates leap days exactly as `lubridate::leap_year()`
}

