% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModelData.R
\name{transf.Date}
\alias{transf.Date}
\title{Obtain functions to transform Date object to and from scaled numeric}
\usage{
\method{transf}{Date}(object, center, scale = 1)
}
\description{
Obtain functions to transform Date object to and from scaled numeric
}

