% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModelData.R
\name{transf}
\alias{transf}
\title{Transform data in a systematic way}
\usage{
transf(object, ...)
}
\description{
Use foo = stats::make.link(); then foo$linkfun(mu), foo$linkinv(eta)
foo = stats::power works the same way
in the future, could customize following http://stackoverflow.com/questions/15931403/modify-glm-function-to-adopt-user-specified-link-function-in-r
}

