% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_checkBDL}
\alias{wqp_checkBDL}
\title{Check for detection limit behavior
Adds is.bdl column
Converts detection-limit units to MeasureValue units.
Augments DetectionQuantitationLimitMeasure.MeasureValue, DetectionQuantitationLimitMeasure.MeasureUnitCode
If detection limit not reported, assumes it is at largest non-bdl value for that dataset}
\usage{
wqp_checkBDL(wqpData)
}
\arguments{
\item{wqpData}{data.frame returned by readWQPData}
}
\description{
Check for detection limit behavior
Adds is.bdl column
Converts detection-limit units to MeasureValue units.
Augments DetectionQuantitationLimitMeasure.MeasureValue, DetectionQuantitationLimitMeasure.MeasureUnitCode
If detection limit not reported, assumes it is at largest non-bdl value for that dataset
}

