% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_simplifyConc}
\alias{wqp_simplifyConc}
\title{Simplify wqp concentration data}
\usage{
wqp_simplifyConc(wqpData)
}
\arguments{
\item{wqpData}{processed WQP data result}
}
\description{
Simplify wqp concentration data
}

