% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_simplifyFlow}
\alias{wqp_simplifyFlow}
\title{Simplify wqp flow data}
\usage{
wqp_simplifyFlow(wqpData)
}
\arguments{
\item{wqpData}{processed WQP data result}
}
\description{
Simplify wqp flow data
}

