% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adry}
\alias{adry}
\title{Calculate antecedent dry days}
\usage{
adry(x, thresh = 0.1)
}
\arguments{
\item{x}{Daily time series of rainfall amount.}

\item{thresh}{threshold precipitation amount for determining what is "dry", defaults to 0.1.}
}
\description{
Calculate antecedent dry days
}

