% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLoad.R
\name{calcLoad}
\alias{calcLoad}
\title{Calculate mass flux from concentration and flow}
\usage{
calcLoad(flow, conc, flow.units = "ft3/s", conc.units = "mg/l",
  load.units = "kg/day")
}
\arguments{
\item{flow}{stream discharge (volume per time)}

\item{conc}{constituent concentration (mass per volume)}

\item{flow.units}{options are "ft3/s" (default), "m3/s", and "l/s"}

\item{conc.units}{options are "mg/l" (default), "ug/l", and "ng/l"}

\item{load.units}{options are "kg/day"}
}
\value{
vector of fluxes in specified units
}
\description{
Performs appropriate unit conversion to go from flow and concentration to load
}

