% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcLoad.R
\name{calcLoad}
\alias{calcLoad}
\title{Calculate loads from concentration and flow}
\usage{
calcLoad(flow, conc, flow.units = c("CFS", "m3/s", "l/s"),
  conc.units = c("mg/l", "ug/l", "ng/l"), load.units = c("kg/day", "kg/yr",
  "tonne/yr"))
}
\arguments{
\item{flow}{stream discharge (volume per timie)}

\item{conc}{constituent concentration (mass per volume)}

\item{flow.units}{options are "CFS" (default), "m3/s", and "l/s"}

\item{conc.units}{options are "mg/l" (default), "ug/l", and "ng/l"}

\item{load.units}{options are "kg/day"}
}
\description{
Performs appropriate unit conversion to go from flow and concentration to load
}

