% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcgam.R
\name{predict.rcgam}
\alias{predict.rcgam}
\title{Predict method for rcgam fits}
\usage{
\method{predict}{rcgam}(object, newdata, flowcol = "flow", smear = TRUE,
  retransform = TRUE, what = c("conc_mg.l", "load_kg.d"), ...)
}
\arguments{
\item{object}{An rcgam object to use for predicting}

\item{newdata}{a data.frame containing precictor variables to use for prediction}

\item{smear}{Use Smearing estimator to correct transformation bias?}

\item{retransform}{Should the predictions be returned as concentrations? (defaults to TRUE)}

\item{...}{Arguments passed to `predict.gam` function call}
}
\description{
Predict values using an rcgam model object
}

