% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcgam.R
\name{predict.rcgam}
\alias{predict.rcgam}
\title{Predict method for rcgam model objects}
\usage{
\method{predict}{rcgam}(object, smear = TRUE, retransform = TRUE,
  type = "response", restrict = FALSE, what = c("concentration", "load"),
  ...)
}
\arguments{
\item{object}{an object of type rcgam, with which to make predictions}

\item{smear}{Unbias the predictions using the smearing coefficient? Defaults to TRUE}

\item{type}{What to predict. If not "response" (the default), then the following are coerced to be FALSE:
retrans, restrict, smear. See ?predict.gam.}

\item{restrict}{Restrict the range of predictions to those observed in the calibration data? Defaults to FALSE.}

\item{...}{Other arguments passed to mgcv::predict.gam}

\item{retrans}{Retransform the predictions into the original units? Defaults to TRUE}
}
\value{
a list with element "fit" and potentially others,
depending on optional arguments passed. See help("predict.gam", package = "mgcv") for more info
The difference is that here the returned value will always be a list, possibly of length 1.
}
\description{
Predict method for rcgam model objects
}

