% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclm.R
\name{predict.rclm}
\alias{predict.rclm}
\title{Predict method for rclm fits}
\usage{
\method{predict}{rclm}(object, smear = TRUE, retransform = TRUE,
  restrict = FALSE, ...)
}
\arguments{
\item{object}{An rclm object to use for predicting}

\item{smear}{Use Smearing estimator to correct transformation bias?}

\item{...}{Arguments passed to `predict.lm` function call}

\item{newdata}{a data.frame containing precictor variables to use for prediction}
}
\description{
Predict values using an rclm model object
}

