% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rclm.R
\name{rclm}
\alias{rclm}
\title{lm rating-curve model}
\usage{
rclm(formula, data, ..., remove.outliers = TRUE, flowcol = "flow",
  lqbar = attr(data, "lqbar"), lqsd = attr(data, "lqsd"),
  ccol = "concentration", datecol = "Date", logflowvar = "logQ",
  logcvar = "logC", doyvar = "doy", numtimevar = "numTime")
}
\description{
Fits linear models to log-log relationships using `mgcv::lm`,
but incorporates aspects
specific to hydrologic rating curves. Specifically, tracks the geometric mean of
flow and concentration, adjusts for transformation bias, and calculates relevant
fit statistics such as Nash-Sutcliffe efficiency (NSE)
}
\details{
Dots get passed to lm() call.
All logarithms are base e (natural log)
Arguments ending in "var" are names of objects that will be assigned in the function.
logflowvar and logcvar will be shifted to have mean zero. POSSIBLY CHANGE THIS?
}

