% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLoad.R
\name{totLoad}
\alias{totLoad}
\title{Calculate total load from a time series of estimated loads
Uses trapezoid approximation for integral.}
\usage{
totLoad(load, datetime, load.units = c("kg/day", "kg/yr", "tonne/yr"))
}
\arguments{
\item{load}{time series of loads.}
}
\description{
Calculate total load from a time series of estimated loads
Uses trapezoid approximation for integral.
}

