% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqData.R
\name{wqp_checkUnits}
\alias{wqp_checkUnits}
\title{Replaces missing units by
1. Looking up using USGSPCode
2. If value is zero, assigns units as most commonly used units in dataset}
\usage{
wqp_checkUnits(wqpData, convertTo = NULL)
}
\arguments{
\item{wqpData}{data.frame returned by readWQPData}

\item{convertTo}{Optional character vector specifying the units to be converted to.
If more than one is provided, they are tried in the order they are given.
The udunits2 package is used for conversion.}
}
\description{
Replaces missing units by
1. Looking up using USGSPCode
2. If value is zero, assigns units as most commonly used units in dataset
}

